pyproject.toml
[project]
name = "ai-career-advisor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "docx2txt>=0.9",
    "envir>=0.1.0",
    "langchain>=0.3.26",
    "langchain-google-genai>=2.1.7",
    "langchain-groq>=0.3.6",
    "numpy>=2.3.1",
    "pdfplumber>=0.11.7",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "spacy>=3.8.7",
    "streamlit>=1.46.1",
]
